+++ verifier/output.go
@@ -4,7 +4,7 @@ package verifier
 // of a name-string.
 type Verification struct {
 	// BestResult returns a result with the best overal score.
-	BestResult *ResultData
+	BestResult *ResultData `json:"bestResult"`
 	// PreferredResults contains verification for data sources the user needs
 	// to be present if available.
 	PreferredResults []*ResultData `json:"preferredResults,omitempty"`
@@ -28,10 +28,25 @@ type ResultData struct {
 	TaxonID string `json:"taxonId,omitempty"`
 	// MatchedName is a verbatim name-string from the matched result.
 	MatchedName string `json:"matchedName,omitempty"`
+	// MatchedCardinality is a number of elements in a name. 0 - no name at all,
+	// 1 - Uninomial, 2 - Binomial, 3 - Trinomial
+	MatchedCardinality int `json:"matchedCardinality,omitempty"`
 	// MatchedCanonical is a canonical form of a matched name
-	MatchedCanonical string `json:"matchedCanonical,omitempty"`
+	MatchedCanonicalSimple string `json:"matchedCanonicalSimple,omitempty"`
+	// MatchedCanonicalFull is a canonical form of a matched name with ranks
+	// and a hybrid sign for named hybrids
+	MatchedCanonicalFull string `json:"matchedCanonicalFull,omitempty"`
 	// CurrentName is a currently accepted name according to the matched result.
 	CurrentName string `json:"currentName,omitempty"`
+	// CurrentCardinality is a number of elements in a name. 0 - no name at all,
+	// 1 - Uninomial, 2 - Binomial, 3 - Trinomial
+	CurrentCardinality int `json:"currentCardinality,omitempty"`
+	// CurrentCanonical is a canonical form of a current name
+	CurrentCanonicalSimple string `json:"currentCanonicalSimple,omitempty"`
+	// CurrentCanonicalFull is a canonical form of a current name with ranks
+	// and a hybrid sign for named hybrids
+	CurrentCanonicalFull string `json:"currentCanonicalFull,omitempty"`
+	// CurrentName is a currently accepted name according to the matched result.
 	// Synonym is true when the name is not the same as currently accepted.
 	Synonym bool `json:"isSynonym,omitempty"`
 	// ClassificationPath of the matched result.
@@ -39,7 +54,7 @@ type ResultData struct {
 	// ClassificationRank of the matched result.
 	ClassificationRank string `json:"classificationRank,omitempty"`
 	// ClassificationIDs of the matched result.
-	ClassificationIDs string `json:"classificationIDs,omitempty"`
+	ClassificationIDs string `json:"classificationIds,omitempty"`
 	// EditDistance tells how many changes needs to be done to apply fuzzy
 	// match to requested name.
 	EditDistance int `json:"editDistance,omitempty"`